add by mass

+++ src/main/java/com/tdsi/sn/app_moblile_api/Services/Verification.java
@@ -59,11 +59,11 @@ public class Verification {
     public String getPt(int password,int telephone){
         return BigInteger.valueOf(password*telephone).toString();
     }
-    public int getPus(int password,int telephone,BigInteger g){
+    public BigInteger getPus(int password,int telephone,BigInteger g){
         BigInteger gs = g.modPow(hashTelephone(password),BigInteger.valueOf(100001));
         BigInteger c = hashTelephone(telephone).multiply(BigInteger.valueOf(-1));
         BigInteger s = gs.add(c).modPow(pub.getPublicExponent(),pub.getModulus());
-        return s.intValue();
+        return s;
     }
     public boolean authenticaed(BigInteger Pu, BigInteger g, Objet objet) throws Exception {
         BigInteger  a = Pu.modPow(pub.getPublicExponent(),pub.getModulus()).add(hashTelephone(objet.getTelephone()));
